+++ starsky/starsky/clientapp/src/hooks/use-gestures/IHandlers.types.ts
@@ -1,0 +1,16 @@
+export interface IHandlers {
+  onPanStart?: Function;
+  onPanMove?: Function;
+  onSwipeLeft?: Function;
+  onSwipeRight?: Function;
+  onSwipeUp?: Function;
+  onSwipeDown?: Function;
+  onPanEnd?: Function;
+  onSwipeLeftEnd?: Function;
+  onSwipeRightEnd?: Function;
+  onSwipeUpEnd?: Function;
+  onSwipeDownEnd?: Function;
+  onPinchStart?: Function;
+  onPinchChanged?: Function;
+  onPinchEnd?: Function;
+}
